// ui/admin/src/models/agentEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wanaku API
 * Wanaku MCP proxy management API
 * OpenAPI spec version: 0.3.0
 */

export interface AgentEntry {
  address: string;
  /** @nullable */
  cardAddress?: string | null;
  description?: string;
  name: string;
  namespace?: string;
}